style: renaming files with some minor modifications to type hinting.
